requirements.txt
Cheetah==2.4.4
Flask==0.10.1
Jinja2==2.8
Landscape-Client==14.12
MarkupSafe==0.23
PAM==0.4.2
Pillow==3.1.1
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Werkzeug==0.11.3
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
cycler==0.10.0
decorator==3.4.0
emd==1.0
gunicorn==19.4.5
html5lib==0.999
itsdangerous==0.24
jsonpatch==1.3
jsonpointer==1.0
matplotlib==1.5.1
meld3==0.6.10
networkx==1.11
numpy==1.10.4
oauth==1.0.1
pandas==0.17.1
prettytable==0.7.2
protobuf==2.6.1
pyOpenSSL==0.13
pycurl==7.19.3
pyparsing==2.1.0
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-dateutil==2.4.2
python-debian==0.1.21-nmu2ubuntu2
pytz==2015.7
requests==2.2.1
scikit-image==0.11.3
scikit-learn==0.17.1
scipy==0.13.3
six==1.5.2
ssh-import-id==3.21
supervisor==3.0b2
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5

